+++ util/CHANGELOG.md
@@ -5,6 +5,7 @@ Refer to https://keepachangelog.com/en/1.0.0/ for how to maintain changelog<br /
 
 ## [Unreleased]
 
+-   add process name prefix to logging
 -   performance: don't pad terminal in production
 -   added `getEnvVar` cross-platform util for getting an environment variable